Removed previous GameFrame classed and renamed GameExample to GameFrame. AB#455
