+++ src/com/team3/forcemajeure/Main.java
@@ -5,7 +5,7 @@ import com.team3.forcemajeure.jswing.*;
 
 public class Main {
     public static void main(String[] args) {
-       new GameExample();
+       new GameFrame();
 
 
     }
