+++ src/com/team3/forcemajeure/jswing/GameFrame.java
@@ -2,19 +2,43 @@ package com.team3.forcemajeure.jswing;
 
 import com.team3.forcemajeure.util.Audio;
 
-import javax.swing.*;
 import java.awt.*;
-import com.team3.forcemajeure.util.*;
-import java.util.Map;
+import java.awt.event.ActionEvent;
+import java.awt.event.ActionListener;
 
+import javax.swing.*;
 
 public class GameFrame {
-    boolean soundOn = true;
+
+    JFrame window;
+    Container con;
+    JPanel menuPanel, userNamePanel, titleNamePanel, startButtonPanel, mainTextPanel, choiceButtonPanel, playerPanel;
+    JLabel userNameLabel, titleNameLabel, ptLabel, ptLabelNumber, inventoryLabel, inventoryLabelName;
+    Font titleFont = new Font("Times New Roman", Font.PLAIN, 90);
+    Font normalFont = new Font("Times New Roman", Font.PLAIN, 28);
+    JButton soundButton, startButton, choice1, choice2, choice3, choice4;
+    JTextArea mainTextArea;
+    int playerPT, monsterHP, silverRing;
+    TitleScreenHandler tsHandler = new TitleScreenHandler();
+    ChoiceHandler choiceHandler = new ChoiceHandler();
+    String inventory, position, player, previousRoom, currentRoom, mainText, firstChoice, secondChoice, thirdChoice, fourthChoice;
+    Boolean soundOn = true;
+    ImageIcon logo = new ImageIcon("resources/images/island.png");
+    ImageIcon gameMapImage;
+    ImageIcon gameBgImage;
+    JLabel mapLabel = new JLabel();
+    JLabel imageBgLabel = new JLabel();
     Audio audio = Audio.getInstance();
-    boolean gameRunning = true;
-    String inputOne;
-    WriteText write = new WriteText();
-//    private static String commandInput;
+
+
+    //Accessor
+    public String getPlayer() {
+        return player;
+    }
+
+    public void setPlayer(String player) {
+        this.player = player;
+    }
 
     public boolean isSoundOn() {
         return soundOn;
@@ -24,48 +48,78 @@ public class GameFrame {
         this.soundOn = soundOn;
     }
 
-    public String getInputOne() {
-        return inputOne;
+    public String getCurrentRoom() {
+        return currentRoom;
     }
 
-    public void setInputOne(String inputOne) {
-        this.inputOne = inputOne;
+    public void setCurrentRoom(String currentRoom) {
+        this.currentRoom = currentRoom;
     }
 
-//    public static String getCommandInput() {
-//        return commandInput;
-//    }
-//
-//    public static void setCommandInput(String commandInput) {
-//        GameFrame.commandInput = commandInput;
-//    }
+    public String getPreviousRoom() {
+        return previousRoom;
+    }
 
-    public GameFrame(){
+    public void setPreviousRoom(String previousRoom) {
+        this.previousRoom = previousRoom;
+    }
 
-        // Images
-        ImageIcon bgImage = new ImageIcon("resources/images/dock.jpg");
-        Image image = new ImageIcon("resources/images/island.png").getImage();
+    public String getMainText() {
+        return mainText;
+    }
+
+    public void setMainText(String mainText) {
+        this.mainText = mainText;
+    }
+
+    public String getFirstChoice() {
+        return firstChoice;
+    }
+
+    public void setFirstChoice(String firstChoice) {
+        this.firstChoice = firstChoice;
+    }
+
+    public String getSecondChoice() {
+        return secondChoice;
+    }
+
+    public void setSecondChoice(String secondChoice) {
+        this.secondChoice = secondChoice;
+    }
+
+    public String getThirdChoice() {
+        return thirdChoice;
+    }
+
+    public void setThirdChoice(String thirdChoice) {
+        this.thirdChoice = thirdChoice;
+    }
+
+    public String getFourthChoice() {
+        return fourthChoice;
+    }
+
+    public void setFourthChoice(String fourthChoice) {
+        this.fourthChoice = fourthChoice;
+    }
+
+    // Ctor
+    public GameFrame() {
+
+        Color bg = Color.black;
+
+        window = new JFrame();
+        window.setSize(1000, 800);
+        window.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
+        window.getContentPane().setBackground(bg);
+        window.setLayout(null);
+        window.setIconImage(logo.getImage());
+        con = window.getContentPane();
 
 
-
-        // Panels
-        JPanel panelOne = new JPanel();
-        panelOne.setBounds(50,10,400,50);
-
-        JPanel panelTwo = new JPanel();
-        panelTwo.setLayout(new GridLayout());
-        panelTwo.setBounds(550,10,400,50);
-
-        JPanel panelThree = new JPanel();
-        panelThree.setBounds(100,70,800,410);
-
-        JPanel panelFour = new JPanel();
-        panelFour.setBounds(250,500,500,125);
-
-
-        // Buttons
-        JButton soundButton = new JButton("🔈 on/off");
-        soundButton.setBounds(25,100,50,50);
+        soundButton = new JButton("🔈 on/off");
+        soundButton.setBounds(15,7,50,50);
         soundButton.addActionListener(e -> {
             if(isSoundOn()) {
                 System.out.println("Sound Off");
@@ -76,192 +130,629 @@ public class GameFrame {
             }
         });
 
-        JButton restartButton = new JButton("🔁 restart");
-        restartButton.setBounds(25,100,50,50);
-        restartButton.addActionListener(e -> System.out.println("Restarting game"));
 
-        JButton quitButton = new JButton("⛱ quit");
-        quitButton.setBounds(50,100,50,50);
-        quitButton.addActionListener(e -> System.out.println("Quitting game"));
+        menuPanel = new JPanel();
+        menuPanel.setBounds(15,7,200,50);
+        menuPanel.setBackground(bg);
 
-        JButton btnNorth = new JButton();
-        btnNorth.setText("⬆");
-        btnNorth.setVerticalTextPosition(JButton.TOP);
-        btnNorth.setBounds(805,500,50,50);
-        btnNorth.addActionListener(e -> {
-                    write.inputText("go north");
+        userNamePanel = new JPanel();
+        userNamePanel.setBounds(350,250,250,125);
+        userNamePanel.setBackground(bg);
+        userNamePanel.setLayout(new FlowLayout(FlowLayout.CENTER));
 
-                }
-               );
+        titleNamePanel = new JPanel();
+        titleNamePanel.setBounds(175, 100, 600, 150);
+        titleNamePanel.setBackground(bg);
+        titleNameLabel = new JLabel("Force Majeure");
+        titleNameLabel.setForeground(Color.white);
+        titleNameLabel.setFont(titleFont);
 
+        startButtonPanel = new JPanel();
+        startButtonPanel.setBounds(370, 400, 200, 100);
+        startButtonPanel.setBackground(bg);
 
-        JButton btnSouth = new JButton();
-        btnSouth.setText("⬇");
-        btnSouth.setBounds(805,550,50,50);
-        btnSouth.addActionListener(e -> {
-                    write.inputText("go south");
-                }
-        );
+        startButton = new JButton("START");
+        startButton.setBackground(bg);
+        startButton.setForeground(Color.white);
+        startButton.setFont(normalFont);
+        startButton.addActionListener(tsHandler);
+        startButton.setFocusPainted(false);
 
-        JButton btnEast = new JButton();
-        btnEast.setText("➡");
-        btnEast.setBounds(855,530,50,50);
-        btnEast.addActionListener(e -> {
-                    write.inputText("go east");
-
-                }
-        );
-
-        JButton btnWest = new JButton();
-        btnWest.setText("⬅");
-        btnWest.setBounds(755,530,50,50);
-        btnWest.addActionListener(e -> {
-                    write.inputText("go west");
-                }
-        );
-        // input into text and output value in class
-
-        // userInput
-        JButton textButton = new JButton("submit");
+        userNameLabel = new JLabel("Enter username");
+        userNameLabel.setForeground(Color.white);
         JTextField textField = new JTextField();
-        textField.setPreferredSize(new Dimension(300,40));
-        textButton.addActionListener(e -> {
-            if(e.getSource() == textButton){
-                System.out.println(textField.getText());
-                write.inputText(textField.getText());
+        textField.setPreferredSize(new Dimension(200,40));
+        startButton.addActionListener(e -> {
+            if(e.getSource() == startButton){
+                setPlayer(textField.getText());
+
             }
         });
 
+        menuPanel.add(soundButton);
+        userNamePanel.add(userNameLabel);
+        userNamePanel.add(textField);
+        titleNamePanel.add(titleNameLabel);
+        startButtonPanel.add(startButton);
 
-        // Labels
-        JLabel labelOne = new JLabel();
-        labelOne.setText("10/100 pts");
-        labelOne.setFont(new Font("Times New Roman", Font.PLAIN, 20));
+        con.add(menuPanel);
+        con.add(userNamePanel);
+        con.add(titleNamePanel);
+        con.add(startButtonPanel);
 
-        JLabel labelInventory = new JLabel();
-        labelInventory.setText("Inventory");
-        labelInventory.setFont(new Font("Times New Roman", Font.PLAIN, 20));
+        window.setVisible(true);
+    }
 
-        JLabel labelTwo = new JLabel();
-        labelTwo.setLayout(new GridLayout());
-        labelTwo.setFont(new Font("Times New Roman", Font.PLAIN, 40));
-        labelTwo.setIcon(bgImage);
+    //Business Methods
+    public void createGameScreen() {
+        userNamePanel.setVisible(false);
+        titleNamePanel.setVisible(false);
+        startButtonPanel.setVisible(false);
 
-        JLabel inputText = new JLabel();
-        inputText.setText(write.getInputString());
-        inputText.setFont(new Font("Times New Roman", Font.PLAIN, 20));
+        mainTextPanel = new JPanel();
+        mainTextPanel.setBounds(220, 75, 600, 425);
+        mainTextPanel.setBackground(Color.black);
+//        con.add(imageLabel);
+        con.add(mainTextPanel);
+        mainTextArea = new JTextArea(
+                "Oops...the text is not showing.");
+        mainTextArea.setBounds(225, 500, 500, 300);
+        mainTextArea.setBackground(Color.black);
+        mainTextArea.setForeground(Color.white);
+        mainTextArea.setFont(normalFont);
+        mainTextArea.setLineWrap(true);
+        mainTextArea.setWrapStyleWord(true);
+        mainTextArea.setEditable(false);
 
-        JLabel outputText = new JLabel();
-        outputText.setText(write.getOutputString());
-        outputText.setFont(new Font("Times New Roman", Font.PLAIN, 20));
+        mainTextPanel.add(mainTextArea);
 
-        // Frame
-        JFrame frame = new JFrame();
-        frame.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
-        frame.setIconImage(image);
-        frame.setLayout(null);
-        frame.setSize(1000,1000);
-        frame.setResizable(true);
-        frame.setVisible(true);
+        choiceButtonPanel = new JPanel();
+        choiceButtonPanel.setBounds(350, 515, 300, 125);
+        choiceButtonPanel.setBackground(Color.black);
+        choiceButtonPanel.setLayout(new GridLayout(4, 1));
+        con.add(choiceButtonPanel);
+        choice1 = new JButton("Choice 1");
+        choice1.setBackground(Color.black);
+        choice1.setForeground(Color.white);
+        choice1.setFont(normalFont);
+        choice1.setFocusPainted(false);
+        choice1.addActionListener(choiceHandler);
+        choice1.setActionCommand("c1");
+        choiceButtonPanel.add(choice1);
+        choice2 = new JButton("Choice 2");
+        choice2.setBackground(Color.black);
+        choice2.setForeground(Color.white);
+        choice2.setFont(normalFont);
+        choice2.setFocusPainted(false);
+        choice2.addActionListener(choiceHandler);
+        choice2.setActionCommand("c2");
+        choiceButtonPanel.add(choice2);
+        choice3 = new JButton("Choice 3");
+        choice3.setBackground(Color.black);
+        choice3.setForeground(Color.white);
+        choice3.setFont(normalFont);
+        choice3.setFocusPainted(false);
+        choice3.addActionListener(choiceHandler);
+        choice3.setActionCommand("c3");
+        choiceButtonPanel.add(choice3);
+        choice4 = new JButton("Choice 4");
+        choice4.setBackground(Color.black);
+        choice4.setForeground(Color.white);
+        choice4.setFont(normalFont);
+        choice4.setFocusPainted(false);
+        choice4.addActionListener(choiceHandler);
+        choice4.setActionCommand("c4");
+        choiceButtonPanel.add(choice4);
 
+        // choice4.setContentAreaFilled(false); // Disable highlighting on press!!!
 
-        // Adding components together
-        panelOne.add(soundButton);
-        panelOne.add(restartButton);
-        panelOne.add(quitButton);
-        panelTwo.add(labelOne);
-        panelTwo.add(labelInventory);
-        panelThree.add(labelTwo);
-        panelFour.add(textButton);
-        panelFour.add(textField);
-        panelFour.add(inputText);
-        panelFour.add(outputText);
-        frame.add(panelOne);
-        frame.add(panelTwo);
-        frame.add(panelThree);
-        frame.add(btnWest);
-        frame.add(btnNorth);
-        frame.add(btnSouth);
-        frame.add(btnEast);
-        frame.add(panelFour);
+        playerPanel = new JPanel();
+        playerPanel.setBounds(250, 0, 600, 50);
+        playerPanel.setBackground(Color.black);
+        playerPanel.setLayout(new GridLayout(1, 4));
+        con.add(playerPanel);
+        ptLabel = new JLabel("Points:");
+        ptLabel.setFont(normalFont);
+        ptLabel.setForeground(Color.white);
+        playerPanel.add(ptLabel);
+        ptLabelNumber = new JLabel();
+        ptLabelNumber.setFont(normalFont);
+        ptLabelNumber.setForeground(Color.white);
+        playerPanel.add(ptLabelNumber);
+        inventoryLabel = new JLabel("Inventory:");
+        inventoryLabel.setFont(normalFont);
+        inventoryLabel.setForeground(Color.white);
+        inventoryLabel.setBackground(Color.red);
+        playerPanel.add(inventoryLabel);
+        inventoryLabelName = new JLabel();
+        inventoryLabelName.setFont(normalFont);
+        inventoryLabelName.setForeground(Color.white);
+        playerPanel.add(inventoryLabelName);
+        playerSetup();
 
-//        frame.pack();
 
     }
 
-    public void createGameScreen(){
+    public void playerSetup() {
+//        audio.play("start");
+        // playerPT = player.getTotalPoints();
+        monsterHP = 20;
+        inventory = "Map";
+        inventoryLabelName.setText(inventory);
+        ptLabelNumber.setText("" + playerPT);
 
+        dock();
     }
 
-    public void runGame() {
-//        Sets up all objects for the game
-        Map<String, Room> roomMap = Data.getRoomMap();
-        Map<String, NPC> npcMap = Data.getNpcMap();
-        Map<String, Endings> endingsMap = Data.getEndingMap();
-        // get starting room ("Water")
-        Room startRoom = roomMap.get("WaterWreckage");
-        // init com.team3.forcemajeure.util.Player
-        Player player = new Player();
-        // set player's current room to start room
-        player.setCurRoom(startRoom);
-//        TODO put enemy in an array that we loop through for encounters
-//        TODO enemies are created from a JSON
-//      Enemy Creation
-        Enemy gambler = new Enemy("gambler", 10, roomMap.get("WaitingRoom"),
-                "You thought you could just get the boatkey, Ha! lets roll for it with dice",
-                "Come back when you are feeling lucky",
-                "I am so embarassed I lost",
-                "boatkey", player);
-
-        while (gameRunning) {
-//            check if ending game condition have been met
-            if (player.getPlayerInventory().getInventory().contains("endgame")){
-                Data.clearInventory();
+    public ImageIcon setImage(String roomName, boolean isMap){
+        ImageIcon anImage = new ImageIcon();
+        //isMap then set map to image else set bg of room to image
+        switch (roomName){
+            case "dock":
+                // show dock image
+                anImage = isMap ? new ImageIcon("resources/images/map/VisitDock/All_Beach.png") : new ImageIcon("resources/images/dock.jpg") ;
                 break;
-            }
-            if (player.getPlayerInventory().getInventory().contains("loopgame")){
-                Data.clearInventory();
-                runGame();
-            }
+            case "talkInstructor":
+                // show talkInstructor image
+                anImage =  isMap ? new ImageIcon("resources/images/map/VisitDock/All_DFl.jpg") : new ImageIcon("resources/images/beach.jpeg");
+                break;
+            case "lobby":
+                // show lobby image
+                anImage =  isMap ? new ImageIcon("resources/images/map/VisitDock/All_DLobby.jpg") : new ImageIcon("resources/images/lobby.jpg");
+                break;
+            case "hall":
+                // show hall image
+                anImage =  isMap ? new ImageIcon("resources/images/map/VisitDock/All_Beach.png") : new ImageIcon("resources/images/hall.jpg");
+                break;
+            case "restaurant":
+                // show restaurant image
+                anImage =  isMap ? new ImageIcon("resources/images/map/VisitDock/All_Beach.png") : new ImageIcon("resources/images/restaurant.jpg");
+                break;
+            case "gameFloor":
+                // show theater image
+                anImage = isMap ? new ImageIcon("resources/images/map/VisitDock/All_Beach.png") :  new ImageIcon("resources/images/casinofloor.jpg");
+                break;
+            case "theater":
+                // show theater image
+                anImage =  isMap ? new ImageIcon("resources/images/map/VisitDock/All_Beach.png") : new ImageIcon("resources/images/theater.jpg");
+                break;
+        }
+        return anImage;
+    }
 
-            // get current room
-            Room curRoom = player.getCurRoom();
-            // play room audio
-            audio.play(curRoom.getName().toLowerCase());
-            // display room info
-            curRoom.displayRoomInfo();
-//            prints story element for doctors office
-            if ((!player.getPlayerInventory().getInventory().contains("larson") &&
-                    !player.getPlayerInventory().getInventory().contains("karma"))){
-                System.out.println(player.getCurRoom().getStory());
-            }
-//            checks if enemy is around
-//            TODO loop through enemy of array when that is implemented
-//            if (player.getCurRoom().equals(gambler.getLocation())) {
-//                gambler.talk();
-//                if (gambler.getHealth() > 0) {
-//                    Battle.battle(gambler, player, 3);
-//                }
-//            }
-//            prompt User
-            String response = TextParser.gameScannerInput();
-            // response needs to be from actionlistenr
-            String response2 = write.getInputString();
-            clrScreen();
-            if ("mute".equals(response2)) {
-                audio.toggleMute();
-            } else {
-                TextParser.gameScannerOutput(response2, player, roomMap, npcMap, endingsMap);
-            }
+    public void setTexts(String pos, String mainText, String choiceOne, String choiceTwo, String choiceThree,
+                         String choiceFour) {
+
+        mainTextPanel.setVisible(true);
+        choice1.setVisible(true);
+        choice2.setVisible(true);
+        choice3.setVisible(true);
+        mapLabel.setVisible(false);
+        imageBgLabel.setVisible(true);
+        position = pos;
+
+        //if room is null then set bgImage to dock else
+        //get bg of image based on room
+        if(getPreviousRoom() == null){
+            gameBgImage = setImage("dock", false);
+            setPreviousRoom("dock");
+
+        } else {
+            gameBgImage = setImage(pos, false);
+            setPreviousRoom(getCurrentRoom());
+        }
+        imageBgLabel.setIcon(gameBgImage);
+        mainTextPanel.add(imageBgLabel);
+        mainTextArea.setText(mainText);
+        choice1.setText(choiceOne);
+        choice2.setText(choiceTwo);
+        choice3.setText(choiceThree);
+        choice4.setText(choiceFour);
+        setMainText(mainText);
+        setFirstChoice(choiceOne);
+        setSecondChoice(choiceTwo);
+        setThirdChoice(choiceThree);
+        setFourthChoice(choiceFour);
+
+        // prev. room = current room
+        // current room = pos
+        setCurrentRoom(pos);
+
+        System.out.println("Prev room: "+ getPreviousRoom() + "\nCurrent room: " + getCurrentRoom());
+    }
+
+
+    // blackjack logic => int
+
+    public void blackJackRound1() {
+        // pos = blackjack
+        // int card = blackjackGame();
+
+        // setText("blackjack", "dealer msg", "hit me", "hold", "","",)
+    }
+
+    public void blackJackRound2() {
+
+    }
+
+    public void blackJackRound3() {
+
+    }
+
+    public void blackJackRound4() {
+
+    }
+
+    public void blackJackRound5() {
+
+    }
+
+
+
+    public void showMap(String currentRoomName){
+        // when clicked => shows map of current room then on choice, return to previous room.
+        setCurrentRoom(currentRoomName);
+        imageBgLabel.setVisible(false);
+        mapLabel.setVisible(true);
+
+        gameMapImage = setImage(getCurrentRoom(), true);
+
+        choiceButtonPanel.setOpaque(true);
+
+        mapLabel.setIcon(gameMapImage);
+        mainTextPanel.add(mapLabel);
+        position = "map";
+        mainTextArea.setText("");
+
+        choice1.setVisible(false);
+        choice2.setVisible(false);
+        choice3.setVisible(false);
+        choice4.setText("exit map");
+    }
+
+
+
+    public void talkInstructor() {
+        setTexts("talkInstructor", "Instructor: Hello " + getPlayer() + ", Please go to the casino and explore", "Go to dock",
+                "Go to lobby", "", "see map");
+    }
+
+    public void beach() {
+        setTexts("beach", "This is the beach", "go to lobby", "go to dock", "", "see map");
+    }
+    public void dock() {
+        setTexts("dock", "This is the dock", "go to beach", "", "", "see map");
+    }
+    public void lobby() {
+        setTexts("lobby", "This is the lobby", "go to hall", "go to beach", "", "see map");
+    }
+    public void hall() {
+        setTexts("hall", "This is the hall", "go to lobby", "go to restaurant", "", "see map");
+    }
+    public void restaurant() {
+        setTexts("restaurant", "This is the restaurant", "go to hall", "go game floor", "", "see map");
+    }
+    public void gameFloor() {
+        setTexts("gameFloor", "This is the game floor", "go to theater", "go to restaurant", "", "see map");
+    }
+    public void theater() {
+        setTexts("theater", "This is the theater", "go to game floor", "", "", "see map");
+    }
+
+    public void ending() {
+
+    }
+    // public void attackGuard(){
+    // position = "attackGuard";
+    // mainTextArea.setText("Guard: Hey don't be stupid!\n\nThe guard fought back
+    // and hit you hard.\n(You receive 3 damage)");
+    // //playerHP = playerHP -3;
+    // playerPT -=3;
+    // ptLabelNumber.setText(""+playerPT);
+    // choice1.setText(">hit me");
+    // choice2.setText(">skip");
+    // choice3.setText("");
+    // choice4.setText("");
+    // }
+    // public void crossRoad(){
+    // position = "crossRoad";
+    // mainTextArea.setText("You are at a crossroad.\nIf you go south, you will go
+    // back to the town.");
+    // choice1.setText("Go north");
+    // choice2.setText("Go east");
+    // choice3.setText("Go south");
+    // choice4.setText("Go west");
+    // }
+    // public void north(){
+    // position = "north";
+    // mainTextArea.setText("There is a river. \nYou drink the water and rest at the
+    // riverside. \n\n(Your HP is recovered by 2)");
+    // playerPT = playerPT + 2;
+    // ptLabelNumber.setText(""+playerPT);
+    // choice1.setText("Go south");
+    // choice2.setText("");
+    // choice3.setText("");
+    // choice4.setText("");
+    // }
+    // public void east(){
+    // position = "east";
+    // mainTextArea.setText("You walked into a forest and found a Long
+    // Sword!\n\n(You obtained a Long Sword)");
+    // inventory = "Long Sword";
+    // inventoryLabelName.setText(inventory);
+    // choice1.setText("Go west");
+    // choice2.setText("");
+    // choice3.setText("");
+    // choice4.setText("");
+
+    // }
+    // public void west(){
+    // position = "west";
+    // mainTextArea.setText("You encounter a goblin!");
+    // choice1.setText("Fight");
+    // choice2.setText("Run");
+    // choice3.setText("");
+    // choice4.setText("");
+    // }
+    // public void fight(){
+    // position = "fight";
+    // mainTextArea.setText("Monter HP: " + monsterHP + "\n\nWhat do you do?");
+    // choice1.setText("Attack");
+    // choice2.setText("Run");
+    // choice3.setText("");
+    // choice4.setText("");
+    // }
+    // public void playerAttack(){
+    // position = "playerAttack";
+
+    // int playerDamage = 0;
+
+    // if(inventory.equals("Knife")){
+    // playerDamage = new java.util.Random().nextInt(3);
+    // }
+    // else if(inventory.equals("Long Sword")){
+    // playerDamage = new java.util.Random().nextInt(12);
+    // }
+
+    // mainTextArea.setText("You attacked the monster and gave " + playerDamage + "
+    // damage!");
+
+    // monsterHP = monsterHP - playerDamage;
+
+    // choice1.setText(">");
+    // choice2.setText("");
+    // choice3.setText("");
+    // choice4.setText("");
+    // }
+    // public void monsterAttack(){
+    // position = "monsterAttack";
+
+    // int monsterDamage = 0;
+
+    // monsterDamage = new java.util.Random().nextInt(6);
+
+    // mainTextArea.setText("The monster attacked you and gave " + monsterDamage + "
+    // damage!");
+
+    // playerPT = playerPT - monsterDamage;
+    // ptLabelNumber.setText(""+playerPT);
+
+    // choice1.setText(">");
+    // choice2.setText("");
+    // choice3.setText("");
+    // choice4.setText("");
+    // }
+    // public void win(){
+    // position = "win";
+
+    // mainTextArea.setText("You defeated the monster!\nThe monster dropped a
+    // ring!\n\n(You obtained a Silver Ring)");
+
+    // silverRing = 1;
+
+    // choice1.setText("Go east");
+    // choice2.setText("");
+    // choice3.setText("");
+    // choice4.setText("");
+
+    // }
+    // public void lose(){
+    // position = "lose";
+
+    // mainTextArea.setText("You are dead!\n\nGAME OVER");
+
+    // choice1.setText("");
+    // choice2.setText("");
+    // choice3.setText("");
+    // choice4.setText("");
+    // choice1.setVisible(false);
+    // choice2.setVisible(false);
+    // choice3.setVisible(false);
+    // choice4.setVisible(false);
+    // }
+    // public void ending(){
+    // position = "ending";
+
+    // mainTextArea.setText("Guard: Oh you killed that goblin!?\nThank you so much.
+    // You are true hero!\nWelcome to our town!\n\nTHE END");
+
+    // choice1.setText("");
+    // choice2.setText("");
+    // choice3.setText("");
+    // choice4.setText("");
+    // choice1.setVisible(false);
+    // choice2.setVisible(false);
+    // choice3.setVisible(false);
+    // choice4.setVisible(false);
+    // }
+
+    public class TitleScreenHandler implements ActionListener {
+
+        public void actionPerformed(ActionEvent event) {
+
+            createGameScreen();
         }
     }
 
-    public void clrScreen(){
-        //TODO DS CLEAR SCREEN FUNCTION
-        for (int i = 0; i < 60; ++i){
-            System.out.println();
+    public class ChoiceHandler implements ActionListener {
+
+        public void actionPerformed(ActionEvent event) {
+
+            String yourChoice = event.getActionCommand();
+
+            switch (position) {
+                case "dock":
+                    switch (yourChoice) {
+                        case "c1":
+                            talkInstructor();
+                            break;
+                        case "c4":
+                            showMap("dock");
+                            break;
+                    }
+                    break;
+                case "map":
+                    switch (yourChoice){
+                        case "c4": setTexts(getCurrentRoom(), getMainText(), getFirstChoice(),getSecondChoice(), getThirdChoice(), getFourthChoice());// get previous method of scene
+                            break;
+                    }
+                    break;
+                case "talkInstructor":
+                    switch (yourChoice) {
+                        case "c1":
+                            dock();
+                            break;
+                        case "c2":
+                            lobby();
+                            break;
+                        case "c4":
+                            showMap("talkInstructor");
+                            break;
+                    }
+                    break;
+                case "lobby":
+                    switch (yourChoice) {
+                        case "c1":
+                            hall();
+                            break;
+                        case "c2":
+                            talkInstructor();
+                            break;
+                        case "c4":
+                            showMap("lobby");
+                            break;
+                    }
+                    break;
+                case "hall":
+                    switch (yourChoice) {
+                        case "c1":
+                            lobby();
+                            break;
+                        case "c2":
+                            restaurant();
+                            break;
+                        case "c4":
+                            showMap("hall");
+                            break;
+                    }
+                    break;
+                case "restaurant":
+                    switch (yourChoice) {
+                        case "c1":
+                            hall();
+                            break;
+                        case "c2":
+                            gameFloor();
+                            break;
+                        case "c4":
+                            showMap("restaurant");
+                            break;
+                    }
+                    break;
+                case "gameFloor":
+                    switch (yourChoice) {
+                        case "c1":
+                            theater();
+                            break;
+                        case "c2":
+                            restaurant();
+                            break;
+                        case "c4":
+                            showMap("gameFloor");
+                            break;
+                    }
+                    break;
+                case "theater":
+                    switch (yourChoice) {
+                        case "c1":
+                            gameFloor();
+                            break;
+                        case "c4": showMap("theater");
+                            break;
+                    }
+                    break;
+                // case "crossRoad":
+                //     switch(yourChoice){
+                //         case "c1": north(); break;
+                //         case "c2": east();break;
+                //         case "c3": dock(); break;
+                //         case "c4": west();break;
+                //     }
+                //     break;
+                // case "north":
+                //     switch(yourChoice){
+                //         case "c1": crossRoad(); break;
+                //     }
+                //     break;
+                // case "east":
+                //     switch(yourChoice){
+                //         case "c1": crossRoad(); break;
+                //     }
+                //     break;
+                // case "west":
+                //     switch(yourChoice){
+                //         case "c1": fight(); break;
+                //         case "c2": crossRoad(); break;
+                //     }
+                //     break;
+                // case "fight":
+                //     switch(yourChoice){
+                //         case "c1": playerAttack();break;
+                //         case "c2": crossRoad(); break;
+                //     }
+                //     break;
+                // case "playerAttack":
+                //     switch(yourChoice){
+                //         case "c1":
+                //             if(monsterHP <1 ){
+                //                 win();
+                //             }
+                //             else{
+                //                 monsterAttack();
+                //             }
+                //             break;
+                //     }
+                //     break;
+                // case "monsterAttack":
+                //     switch(yourChoice){
+                //         case "c1":
+                //             if(playerPT <1 ){
+                //                 lose();
+                //             }
+                //             else{
+                //                 fight();
+                //             }
+                //             break;
+                //     }
+                //     break;
+                // case "win":
+                //     switch(yourChoice){
+                //         case "c1": crossRoad();
+                //     }
+                //     break;
+
+            }
         }
     }
-
-
 }
+
